+++ controllers/index.js
@@ -5,5 +5,5 @@ const homeRoutes = require('./homeRoutes');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
-
+router.use('upload', uploadRoutes)
 module.exports = router;
