+++ server.js
@@ -4,6 +4,7 @@ const session = require('express-session');
 const exphbs = require('express-handlebars');
 const routes = require('./controllers');
 const helpers = require('./utils/helpers');
+
 const uploader = require('./utils/upload');
 
 const sequelize = require('./config/connection');
@@ -50,20 +51,18 @@ if (typeof (process.env.CLOUDINARY_URL) === 'undefined') {
 }
 
 
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, 'uploads/')
-  },
-  filename: function (req, file, cb) {
-    cb(null, file.originalname)
-  }
-});
+// const storage = multer.diskStorage({
+//   destination: function (req, file, cb) {
+//     cb(null, 'uploads/')
+//   },
+//   filename: function (req, file, cb) {
+//     cb(null, file.originalname)
+//   }
+// });
 
-const upload = multer({ storage });
+// const upload = multer({ storage });
 
-
-
-app.post('/upload', upload.fields([{ name: 'file' }]), uploader.uploadImage);
+// app.post('/upload', upload.fields([{ name: 'file' }]), uploader.uploadImage);
 
 // Inform Express.js on which template engine to use
 app.engine('handlebars', hbs.engine);
