+++ utils/upload.js
@@ -1,27 +1,24 @@
 const cloudinary = require('cloudinary').v2;
-const { Router } = require('express');
-const multer = require('multer');
-const router = require('express').Router();
 const path = require('path');
 const applicationDirectory = path.dirname(require.main.filename);
 console.log("applicationDirectory", applicationDirectory)
-
+const util = require('node:util');
 
 //cloudinary returns a promise
 //this function returns a URL string 
-const uploadImage = async filename => {
-    filename = `${applicationDirectory}/uploads/${filename}`;
+const uploadImage = async (req, res) => {
+    const uploadedFile = req.files.file[0];
+    const filename = uploadedFile.filename;
+    pathname = `${applicationDirectory}/uploads/${filename}`;
     try {
-        const photo = await cloudinary.uploader.upload(filename, {
+        const photo = await cloudinary.uploader.upload(pathname, {
             use_filename: true,
             unique_filename: false
         });
         console.log('after cloudinary upload', photo)
 
-        //return the "raw" asset (uncomment the return statement below)
-        // return photo;
-
-        // return transformed img (uncomment the return statement below)
+        
+        res.status(200).JSON(photo);
         return await cloudinary.url(photo.public_id, {
             width: 400,
             quality: 'auto',
@@ -36,5 +33,4 @@ const uploadImage = async filename => {
 };
 
 
-module.exports = {uploadImage};
-
+module.exports = { uploadImage };
